Enforce paid amounts must not be present if case status is not closed
Adds a dataframe check to the schema that fails if paid amounts
are present when case status is open. Adds an additional optional cleaning script
to remove paid amounts when case status is open.

Also contains a few minor fixes to individual states' converter yaml files
and cleaning scripts. These are not related to the paid amounts issue.