Add custom reports for each state
This commit adds 2 new tasks to the pipeline:
    1. 5-get-report-data - preproceses the data for the reports
    2. 6-generate-state-reports - generates a report for each state

The reports are generated using a jinja2 template. The template is in the hand/ directory.

The reports are generated in the tasks/6-generate-state-reports/output/ directory.

Also adds a new function to the shared_functions.py file to add missing texas amounts data. That function was originally in the notebooks/national_analysis/national_analysis.ipynb file.